+++ util/cassert.c
@@ -21,7 +21,7 @@ void assert_bytes_eq(size_t expectedlen, const uint8_t *expected, \
 		     const uint8_t *actual)
 {
 	int i;
-	for (i = 0; i < expectedlen; ++i)
+	for (i = 0; (size_t)i < expectedlen; ++i)
 		assert(actual[i] == expected[i]);
 }
 
@@ -29,6 +29,6 @@ void assert_farrs_eq(size_t expectedlen, const float *expected,
 		     const float *actual, float eps)
 {
 	int i;
-	for (i = 0; i < expectedlen; ++i)
+	for (i = 0; (size_t)i < expectedlen; ++i)
 		assert_floats_eq(expected[i], actual[i], eps);
 }
